template: Use PRIuPTR to output uintptr_t segsize
This is a manual rewrite "cherry-pick" of:

68360d49216cb9ddda62ee6dbe30663e8dbd23f0
template: Use PRIuPTR to output uintptr_t segsize/minheapoffset

+++ ibv-conduit/gasnet_core.c
@@ -2108,8 +2108,8 @@ extern int gasnetc_attach( gasnetex_client_t      *client_p,
                            int                    numentries,
                            uintptr_t              segsize)
 {
-  GASNETI_TRACE_PRINTF(C,("gasnetc_attach(table (%i entries), segsize=%lu)",
-                          numentries, (unsigned long)segsize));
+  GASNETI_TRACE_PRINTF(C,("gasnetc_attach(table (%i entries), segsize=%"PRIuPTR")",
+                          numentries, segsize));
 
   if (!gasneti_init_done) 
     GASNETI_RETURN_ERRR(NOT_INIT, "GASNet attach called before init");
